Refactor and extract CanonicalPath into a common header
CanonicalPath class and the node types are a common type which when split
from the hir-mapping class decouples the headers and cleans things up.
