+++ gcc/rust/resolve/rust-name-resolver.h
@@ -20,117 +20,12 @@
 #define RUST_NAME_RESOLVER_H
 
 #include "rust-system.h"
+#include "rust-canonical-path.h"
 #include "rust-hir-map.h"
 #include "rust-hir-type-check.h"
 
 namespace Rust {
 namespace Resolver {
-
-// https://doc.rust-lang.org/reference/paths.html#canonical-paths
-//
-// struct X - path X
-// impl X { fn test - path X::test }
-//
-// struct X<T> - path X
-//
-// impl X<T>   { fn test - path X::test}
-// impl X<i32> { fn test - path X<i32>::test }
-// impl X<f32> { fn test - path X<f32>::test }
-//
-// pub trait Trait { // ::a::Trait
-//   fn f(&self); // ::a::Trait::f
-// }
-//
-// impl Trait for Struct {
-//    fn f(&self) {} // <::a::Struct as ::a::Trait>::f
-// }
-class CanonicalPath
-{
-public:
-  CanonicalPath (const CanonicalPath &other) : segs (other.segs) {}
-
-  CanonicalPath &operator= (const CanonicalPath &other)
-  {
-    segs = other.segs;
-    return *this;
-  }
-
-  static CanonicalPath new_seg (NodeId id, const std::string &path)
-  {
-    rust_assert (!path.empty ());
-    return CanonicalPath ({std::pair<NodeId, std::string> (id, path)});
-  }
-
-  std::string get () const
-  {
-    std::string buf;
-    for (size_t i = 0; i < segs.size (); i++)
-      {
-	bool have_more = (i + 1) < segs.size ();
-	const std::string &seg = segs.at (i).second;
-	buf += seg + (have_more ? "::" : "");
-      }
-    return buf;
-  }
-
-  static CanonicalPath get_big_self (NodeId id)
-  {
-    return CanonicalPath::new_seg (id, "Self");
-  }
-
-  static CanonicalPath create_empty () { return CanonicalPath ({}); }
-
-  bool is_empty () const { return segs.size () == 0; }
-
-  CanonicalPath append (const CanonicalPath &other) const
-  {
-    rust_assert (!other.is_empty ());
-    if (is_empty ())
-      return CanonicalPath (other.segs);
-
-    std::vector<std::pair<NodeId, std::string>> copy (segs);
-    for (auto &s : other.segs)
-      copy.push_back (s);
-
-    return CanonicalPath (copy);
-  }
-
-  // if we have the path A::B::C this will give a callback for each segment
-  // example:
-  //   A
-  //   A::B
-  //   A::B::C
-  void iterate (std::function<bool (const CanonicalPath &)> cb) const
-  {
-    std::vector<std::pair<NodeId, std::string>> buf;
-    for (auto &seg : segs)
-      {
-	buf.push_back (seg);
-	if (!cb (CanonicalPath (buf)))
-	  return;
-      }
-  }
-
-  NodeId get_id () const
-  {
-    rust_assert (!segs.empty ());
-    return segs.back ().first;
-  }
-
-  bool operator== (const CanonicalPath &b) const
-  {
-    return get ().compare (b.get ()) == 0;
-  }
-
-  bool operator< (const CanonicalPath &b) const { return get () < b.get (); }
-
-private:
-  explicit CanonicalPath (std::vector<std::pair<NodeId, std::string>> path)
-    : segs (path)
-  {}
-
-  std::vector<std::pair<NodeId, std::string>> segs;
-};
 
 class Rib
 {
